Add RSA algorithm with archive signing support
Add Archive::Tyd::Algorithm::RSA, which (so far) only supports code
signing for Tyd archives. Add support for custom_block(), which so far
is only useful for storing/retrieving signatures. Add file_table() which
returns a canonical file table string for signing purposes. Add verify()
which verifies the signature on an archive and ensures all the member
files are intact.