Add support for RSA and file table encryption
Add support to Archive::Tyd::Algorithm::RSA to use RSA public key
encryption on the contents of files (in addition to code signing). Add
the ability for the file table itself to also be encrypted using
whatever encryption algorithm you want (so that the resulting archive
includes an encrypted [file-table] block, so that not even the file
names and attributes are exposed without the correct decryption key).